# Fix perf issue in System.Diagnostics.Activity